# openapi/v0.yaml
openapi: 3.1.0
info:
  title: Fluorite Registry API
  version: "0.1.0"
  description: >
    Registry for publishing and fetching compiled TurboWarp extensions built
    with the Fluorite Compiler. All endpoints are versioned under /v0.
servers:
  - url: https://fluorite.sdisk.us/v0
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Accepts two kinds of bearer token: session tokens (from /auth/login,
        full access, 7-day TTL) and automation tokens (from /auth/tokens,
        long-lived, restricted to scopes granted at creation — currently only
        "publish"). Scope enforcement happens server-side per-endpoint.
  parameters:
    namespaceParam:
      name: namespace
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$"
    extensionIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9](?:[a-zA-Z0-9._-]{0,63})$"
    versionParam:
      name: version
      in: path
      required: true
      schema:
        type: string
        description: A literal semver string, or "latest".
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated, but not permitted to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VERSION_PENDING_REVIEW
            message:
              type: string
              example: >-
                You have another version pending review. It needs to be
                approved or rejected before you can publish again.
            field:
              type: [string, "null"]
...
    VersionSummary:
      type: object
      required: [version, status, createdAt, yanked, downloads]
      properties:
        version:
          type: string
        status:
          type: string
          enum: [staging, pending, published, rejected]
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: [string, "null"]
          format: date-time
        yanked:
          type: boolean
          default: false
        yankReason:
          type: [string, "null"]
        downloads:
          type: integer
          description: "Count of compiled-code fetches (Accept: application/javascript), incremented per request."
    VersionMetadata:
      allOf:
        - $ref: "#/components/schemas/VersionSummary"
        - type: object
          required: [id, name, license, description]
          properties:
            id:
              type: string
            name:
              type: string
            license:
              type: string
            description:
              type: string
...
    Webhook:
      type: object
      required: [id, url, events, enabled, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - version.published
              - version.pending
              - version.approved
              - version.rejected
              - version.yanked
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        secret:
          type: string
          description: >-
            Only present in the response to the create call. Used to compute
            the X-Fluorite-Signature HMAC header on delivered payloads; not
            retrievable afterward.
...
paths:
...
  /extensions/@{namespace}/{id}/versions:
    post:
      tags: [Userspace]
      operationId: publishVersion
      summary: >-
        Publish a new version. The manifest (id, name, version, license,
        description) is extracted from the extension source itself — there
        is no separate metadata body. Accepts session tokens or automation
        tokens scoped to "publish".
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
      requestBody:
        required: true
        content:
          application/javascript:
            schema:
              type: string
      responses:
        "201":
          description: Version created (pending review, or published if the owner is trusted). Fires version.pending or version.published webhooks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionMetadata"
        "403":
          description: Another version is already pending review for this owner, or the token's scope doesn't permit publishing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: This exact version already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /extensions/@{namespace}/{id}/versions/{version}:
    get:
      tags: [Open]
      operationId: getVersion
      summary: >-
        Get a version. Content negotiation determines the response:
        application/json for metadata, application/javascript for the
        compiled code. Fetching the compiled code increments its download
        counter; fetching metadata does not. Yanked versions are still
        servable (existing consumers keep working) but are flagged.
      security: []
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      responses:
        "200":
          description: Version metadata or compiled code, depending on Accept.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionMetadata"
            application/javascript:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Adminspace]
      operationId: updateVersion
      summary: Approve or reject a pending version. Fires version.approved or version.rejected webhooks.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                reason:
                  type: string
      responses:
        "200":
          description: Status updated. A notification is sent to the owner in the same transaction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionMetadata"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Userspace]
      operationId: deleteVersion
      summary: Remove a single version (and its blob) without touching the rest of the extension. Owner or admin.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
        - $ref: "#/components/parameters/extensionIdParam"
        - $ref: "#/components/parameters/versionParam"
      responses:
        "204":
          description: Deleted.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /versions:
    get:
      tags: [Adminspace]
      operationId: listVersionsAcrossOwners
      summary: >-
        A cross-owner, filterable view of versions. This is the moderation
        worklist: filter by status=pending to find everything awaiting
        review, without already knowing which namespace/id it belongs to.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [staging, pending, published, rejected]
        - $ref: "#/components/parameters/cursorParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A page of versions, each including its owner namespace and extension id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/VersionMetadata"
                        - type: object
                          required: [namespace, extensionId]
                          properties:
                            namespace:
                              type: string
                            extensionId:
                              type: string
                  nextCursor:
                    type: [string, "null"]
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /webhooks:
    get:
      tags: [Adminspace]
      operationId: listWebhooks
      summary: List configured webhooks.
      responses:
        "200":
          description: Webhooks. Secrets are not included.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      tags: [Adminspace]
      operationId: createWebhook
      summary: Register a webhook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    enum:
                      - version.published
                      - version.pending
                      - version.approved
                      - version.rejected
                      - version.yanked
      responses:
        "201":
          description: Webhook created. The secret is returned once, here only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /webhooks/{id}:
    patch:
      tags: [Adminspace]
      operationId: updateWebhook
      summary: Update a webhook's URL, events, or enabled state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Adminspace]
      operationId: deleteWebhook
      summary: Remove a webhook.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...